Allow building an image using a local Neo4j build
This allows us to test product changes before they are released.

Because the build environment is isolated from the host there doesn't
seem to be any way to expose the development tarball in a way that
curl can grab it directly (via either file:// or http://). So we have
to inject the tarball directly into the image. That makes a bit of a
mess of the non-dev Dockerfiles, but I can't find a better
alternative.