Run up a "dev" container for testing
This is a bit of a hack in order to accommodate the need to interrupt
the container once it's up. It seems that we can trap the SIGINT to stop
Make from marking the target as failed, but it still stops Make from
processing further targets. We therefore put the dependency of "all" on
"dev/runs-okay" at the end of the Makefile, so that the other targets
will run first.