Go back to running as rooot
The solution here does not work when mounting volumes from directories
that are owned by the user running Docker. The best solution currently
involves running the entrypoint script as root, modifying the ownership
of the volume in there and then droppping down to the neo4j user (see
Redis' example,
https://github.com/docker-library/redis/tree/master/3.0). I'm going to
tackle that later.