Completely ignore SIGINT of the container in test
The test runs Neo4j in a container in the foreground to give us a chance
to interact with it if necessary. In order to end the test we need to
interrupt the Neo4j process.

By default this interrupt makes the test shell exit non-zero so Make
treats it as having failed. So we have been trapping the signal in the
recipe. However the parent process (Make) also handles the interrupt;
so, although the target succeeds, subsequent targets to do not run and
Make exits with a non-zero status.

We've worked around this in the past by ordering the targets in the
Makefile such that dev/runs-okay was the last thing to run. However
that is fragile and may make the Makefile harder to read.

This solution does a better job and is not sensitive to the order of the
Makefile. It puts the shell running the container into monitor mode, so
it runs in its own process group and does not propagate signals up to
the parent. Monitor mode is turned off before spawning the docker
subprocess so that the script does get the interrupt from that process.