Make start script more flexible to accept location
Add a variable to describe an alternative to the root location for most
of the extra files:
    /conf
    /ssl
    /plubins
    _extension_script

all the above are prepended with an environment variable
ALTERNATE_DIR_LOCATION. The default value is empty. Wich does not change
previous behavior.

The ALTERNATE_DIR_LOCATION can point to a different location to find
those assets mentioned above, e.g., /my/extracted/zip/dir.

The variable can also be set to either another variable name, e.g.,
$MESOS_SANDBOX, in which case it will be evaluated to the content of the
later variable.

Also on this PR, the ability to set different ports for the cluster and
the server communication. Also quite useful for deployments on Mesos.

Latest, there's an improvement to the _settings_ function in order to be
able to append inexistant configurations.