Allow longer for temporary background database to start when changing password
This start-up was timing out on Debian Jessie but not other
platforms. The timeout should have been longer anyway because, while
it's reasonable to expect the database to start in 10 seconds, we can
afford to give it a very long timeout to accommodate slow edge cases
since it will only fail if something is badly wrong.

Fixes #32.