Calculate the final download URI at build time
At the point where we build the Dockerfile Neo4j hasn't been released,
so the dist.neo4j.org URI doesn't actually exist yet. But we know what
it will be ahead of time, so we can insert it into the Dockerfile
nevertheless. This simplifies the whole pipeline because we don't need
to inject anything at release time.

This change introduces a trick for putting dependencies on environment
variables into Make rules. We write the value of the variable into a
file iff it changes and depend on that. This will enable us to build the
Community and Enterprise images in the same directory without depending
on a `make clean`.