Set settings that aren't already present
The current version of setting() only edits settings already present in the config files. I discovered this while trying to set a read_only setting but it holds true in general. The version included sets appends settings if they are not already present in the config file. If this change is accepted it would have to be applied to all copies of the entrypoint script.

Peter