Allow a local package to be injected by later steps in the build pipeline
When testing the artifacts in later build steps we need to be able to
use a local package rather than the one published to dist.neo4j.org,
because while we are testing the release will not yet have been
published to that server. However we don't want later build steps to
have to render a template to produce the Dockerfile with the correct URI
in it because that requires special knowledge of how the templating
works.

So instead we use a mechanism based on an "optional" COPY in the
Dockerfile (copying a directory that may be empty) and an optional build
argument for specifying the local file's URI. Specifically, to test
against a local package:

 * place the package in the `local-package` directory
 * invoke `docker build` with `--build-arg=file://tmp/<package-filename>`

If you don't pass the build argument then the public distribution URI
that's embedded in the Dockerfile at build time will be used.

Note that we still use the SHA that's embedded in the Dockerfile to
verify that we're using the correct package. This allows us to be
satisfied that we've already tested the Docker image using the correct
Neo4j build even though it hasn't been released yet.