Move 3.1 to use an Alpine-based base image
This is for security reasons: it reduces the attack surface because
Alpine is a "minimal" distro. There are also unfixed security holes in
the Debian base image; the Docker maintainers recommend using Alpine
instead because the upstream reacts quicker to these problems.

We have decided that, although it is a security risk, making the change
in a point release of one of the older release is too much of a change.