Restore Linux compatibility
A previous commit attempted to use a sed invocation that would work on
OSX and Linux. But it doesn't. The wisdom of StackOverflow is that
there is no such invocation if you want to use in-place editing. We
can work around this somehow by not doing in-place, but I'm going to
leave that until someone actually needs to run this on OSX again.