Ensure config only has single entries
This fixes a bug where dbms.mode would be set multiple times in the
config. Turns out that this is fine for all values EXCEPT
dbms.mode=ARBITER.

In that case we rely on parsing the config from bash which runs into a
problem due to:
https://github.com/neo4j/neo4j/blob/6cc2ebb93b39405976801f57173079deea45fd11/packaging/standalone/src/main/distribution/shell-scripts/bin/neo4j-shared.m4#L157

If an entry is defined multiple times, it is regarded as a cumulative
option. This is used for JAVA_OPTS and similarly. But for Arbiter we
set a different main class. So having multiple dmbs.mode entries will
set an incorrect value which means the default main class is called,
instead of the Arbiter's.

So this fix makes sure to only replace uncommented values in the
config, and otherwise always append to the end. Previously we would
replace even commented values, but since dmbs.mode is listed twice
in the 3.1/3.2/3.3 config (once in the HA-section, once in CC-section)
it would get set twice.