Run Neo4j as non-root
This builds on https://github.com/neo4j/docker-neo4j/pull/22

It uses `su-exec`
[instead of gosu](https://github.com/tianon/gosu#su-exec) as
recommended by the author of `gosu` himself for Alpine images.

I have confirmed that docker these days sets a sensible ulimit so this
is no longer an issue we need to consider.

There is a fair amount of special handling in the entrypoint for for
file permissions. We chown the relevant directories to the neo4j user
to make them writable for the process but we take care not to chown
directories which are volumes (to reasonable extents).

If a volume is mounted in `/data` or `/conf` (in that order) the neo4j
user is created with the same UID and GID as those directories. This
has the effect that the container will write new files with the same
UID/GID as the user owning the host directory. So once the container
exists, file will not have surprising owners.

Tests have been added to verify that the UID of files written are
sensible when volume mounts exist.