Switch to AdoptOpenJDK-based JVM base image
The background here is that - again - the Debian OpenJDK builds turn out
to have the wrong version numbers on them, with the package being
stamped as a java update that was released after the package was built.

See https://mail.openjdk.java.net/pipermail/jdk8u-dev/2019-May/009330.html

We don't use the Debian packages, we use Alpine, but inspecting the
provenance of the Alpine OpenKJDK packaging, it seems to suffer from the same
issue. The current latest alpine OpenJDK has versioning that does not
seem to line up. At the very least, the source of those builds is of a
similar kind to Debian.

This moves our base image to images provided by AdoptOpenJDK, a well
regarded source of up-to-date OpenJDK builds. The base image remains
Alpine Linux.