TestMounting was incorrectly passing some data mounting tests, fixed.
We were using a Neo4jContainer from TestContainers rather than a GenericContainer and this
seems to have caused the /data/dbms directory to get built unexpectedly.
The data/dbms folder contains authentication information, and it turns out that Neo4jContainer
was setting a default password even though NEO4J_AUTH=none is set, hence the folder appeared and out tests passed.
I've added an initial password so that the folder is created and we can test permissions on it.