Add open options settings for test configs
Previously some module options were hardcoded in neo4j. Now they are part of distributes neo4j config. As a result, if config gets lost or changed for test purposes we need to make sure we have a minimal set of options that will actually allow neo4j to be started.