Update Debian package manager usage
The man pages of Debian quite clearly states that `apt` is for human
usage, and scripts should rely on `apt-get` for stability.

Also, packages should be installed with `-no-install-recommends` to
avoid pulling in unnecessary dependencies. E.g. in the current builds
`gcc` pulls in `libtiff5` without us really needing it.

No longer installing recommended packages means that we have to
complement our packages with `ca-certificates` and `libc-dev`.